gmail_idealista_app/requirements.txt
google-api-python-client==2.170.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
python-dateutil==2.9.0.post0
